Moved recyclermvp package to separate module.
