+++ recyclermvp/src/main/res/values/strings.xml
@@ -1,0 +1,3 @@
+<resources>
+	<string name="mvp.recycler.view.title">Mvp RecyclerView</string>
+</resources>
